Move components that connect to redux store to container folder
