+++ src/container/Filter.js
@@ -5,12 +5,7 @@ import { filterPokemons } from '../actions/pokeActions';
 import details from '../css/Details.module.css';
 
 const Filter = props => {
-  // const [state, setState] = useState('');
-  // console.log(props);
   const handleChange = e => {
-    // console.log(e.target.value);
-    // setState(e.target.value);
-    // console.log(state);
     // eslint-disable-next-line react/prop-types
     const { filterPokemons } = props;
     filterPokemons(e.target.value);
