+++ src/container/PokeDetails.js
@@ -11,45 +11,16 @@ import details from '../css/Details.module.css';
 import { fetchPokemon } from '../actions/index';
 import { getOnePokemon, getPokemonPending } from '../reducers/pokeReducer';
 
-// const apiUrl = 'https://pokeapi.co/api/v2/pokemon';
 class PokeDetails extends Component {
-  // constructor(props) {
-  //   super(props);
-
-  //   this.state = {
-  //   };
-  // }
-
   componentDidMount() {
-    // this.getDetails();
     const { fetchPokemon } = this.props;
     this.name = this.props.match.params.name;
     fetchPokemon(this.name);
   }
 
-  // getDetails() {
-  //   this.name = this.props.match.params.name;
-  //   fetch(`${apiUrl}/${this.name}`)
-  //     .then(res => res.json())
-  //     .then(resItems => {
-  //       this.show(resItems);
-  //     });
-  // }
-
-  // show(items) {
-  //   this.itemx = items;
-  //   // console.log(this.itemx);
-  //   this.setState({
-  //     ...this.itemx,
-  //   });
-  //   // console.log(this.state);
-  // }
-
   render() {
     const poke = this.props.pokemon;
     const { loading } = this.props;
-    // console.log('My Poke', poke);
-    // console.log(poke.sprites && poke.sprites.other.dream_world.front_default);
     return (
       <>
         {loading && (
